unify bbox_utils
