+++ ppdet/modeling/__init__.py
@@ -7,7 +7,6 @@ from . import losses
 from . import architectures
 from . import post_process
 from . import layers
-from . import utils
 
 from .ops import *
 from .backbones import *
@@ -18,4 +17,3 @@ from .losses import *
 from .architectures import *
 from .post_process import *
 from .layers import *
-from .utils import *
